First Commit
This is a working commit of the python script (mos.py) which will grab
an image from a webcam on /dev/video0 and then add it to a gallery of
images on-screen.
The images pixellated before they’re added to the gallery.

The other scripts and files are left-over from the development of this
tool.