Additional Scripts to test various concepts
facefind.py and findface2.py to carry out face detection (use
findface2.py, which was found on stackoverflow).
filelist.py lists the files in a directory, and loads the relevant ones
for display on a pyame window
haarcascade…xml is facial recognition data required for OpenCV
photobooth.py is where all the building blocks are put together.