Initial working version of the app
Photo Booth.py
————————

This is a ‘working’ version of the app.
It will take an image when the space bar is pressed and then display it
onscreen in a grid with the last few collected images.
It won’t respond to button presses (although this is relatively trivial)
It won’t read/save files to disk yet
It will find faces in an image, but the size of the image returned will
vary depending on the scale of the face.
It won’t find multiple faces in an image
There is no LED control in case illumination is required